#149 Linting and cleaning
